Try to make CI correct.
