--- futures-await/.travis.yml
@@ -1,12 +1,0 @@
-language: rust
-rust: nightly
-sudo: false
-
-script:
-  - cargo test
-  - cargo clean
-  - cargo test --manifest-path testcrate/Cargo.toml
-
-notifications:
-  email:
-    on_success: never
