+++ futures-macro-async/Cargo.toml
@@ -14,9 +14,12 @@ proc-macro = true
 
 [dependencies]
 quote = "0.4"
-proc-macro2 = { version = "0.2", features = ["nightly"] }
+proc-macro2 = "0.2"
 
 [dependencies.syn]
 version = "0.12.13"
 features = ["full", "fold", "parsing", "printing", "extra-traits"]
 default-features = false
+
+[features]
+nightly = ["proc-macro2/nightly"]
