+++ futures/testcrate/Cargo.toml
@@ -6,8 +6,8 @@ authors = ["Alex Crichton <alex@alexcrichton.com>"]
 [lib]
 path = "lib.rs"
 
-[dependencies]
-futures-await = { path = ".." }
+[dependencies.futures]
+path = ".."
 
 [dev-dependencies]
 compiletest_rs = "0.3.7"
@@ -15,3 +15,6 @@ compiletest_rs = "0.3.7"
 [[test]]
 name = "ui"
 harness = false
+
+[features]
+nightly = ["futures/nightly"]
