+++ futures/testcrate/ui/missing-item.rs
@@ -1,7 +1,7 @@
 #![allow(warnings)]
 #![feature(proc_macro, conservative_impl_trait, generators, pin)]
 
-extern crate futures_await as futures;
+extern crate futures;
 
 use futures::prelude::*;
