+++ futures/testcrate/ui/move-captured-variable.stderr
